Resolve #10
This resolved issue #10 for me because it grabs the first project directory. 

I think it's a bit hack-y, what if you have multiple directories in your project? However, probably only the minority of people have that.

I think this is more of a proof of concept than a proposed change, but it might be good enough? Just putting it out there so people can test, comment, and maybe propose a better solution.